requirements.txt
opencv-python
ultralytics
numpy
termcolor
streamlit
pillow
opencv-python-headless
